Levenshtein Distance algorithm in java
Levenshtein Distance algorithm is modified such that result is converted into range between 0-100 percentage. 0-indicates two strings are same. 100-indicates all the characters are different. 

How this change is making difference? 
for eg, When I want to find how much dissimilar the two strings are: 
eg 1: input: str1="it" str2= "on"   Levenshtein Distance algorithm  output:1
eg2: input: str1="example" str2="exampla"     Levenshtein Distance algorithm  output:1

in both examples output is 1. but when you  convert it into percentage, it will be normalized/standardized) and it will be not dependent on length of the string.

Where this change be useful:
When I need to implement hierarchical clustering algorithm, I need the distance function to say how close the two objects are.